Containerized/requirements.txt
fastapi
uvicorn
pillow
numpy
requests
onnxruntime
google-cloud-storage
